async loading and query init
