+++ pkg/interactive/interactive_client.go
@@ -462,7 +462,7 @@ func (c *InteractiveClient) getQuery(ctx context.Context, line string) *modconfi
 		// show the current status of initialization in the spinner
 		statushooks.SetStatus(ctx, c.initData.Status)
 		// set up initdata so that we can update the spinner when the status changes
-		c.initData.InitData.OnStatusChanged = func(newStatus string) {
+		c.initData.OnStatusChanged = func(newStatus string) {
 			statushooks.SetStatus(ctx, newStatus)
 		}
 		// wait for client initialisation to complete
